+++ src/Logic.h
@@ -16,12 +16,12 @@ Logic.h
 
 class Logic {
 	public:
-		Logic(std::string type, std::string name, std::vector<Connector*> inputs, std::vector<Connector*> outputs, bool sign, float delay, std::string outType, int timeASAP, int timeALAP);
+		Logic(std::string type, std::string name, std::vector<Connector*> inputs, std::vector<Connector*> outputs, bool sign, int delay, std::string outType, int timeASAP, int timeALAP);
 		//Get
 		std::string get_type() { return this->type; }
 		std::string get_name() { return this->name; }
 		bool get_sign() { return this->sign; }
-		double get_delay() { return this->delay; }
+		int get_delay() { return this->delay; }
 		std::string get_outType() { return this->outType; }
 		std::vector<Connector*> get_inputs() { return this->inputs; }
 		std::vector<Connector*> get_outputs() { return this->outputs; }
@@ -36,7 +36,7 @@ class Logic {
 		void set_type(std::string type) { this->type = type; }
 		void set_name(std::string name) { this->name = name; }
 		void set_sign(bool sign) { this->sign = sign; }
-		void set_delay(double delay) { this->delay = delay; }
+		void set_delay(int delay) { this->delay = delay; }
 		void set_outType(std::string outType) { this->outType = outType; }
 
 		void set_timeASAP(int timeASAP) { this->timeASAP = timeASAP; }
@@ -51,7 +51,7 @@ class Logic {
 		std::vector<Connector*> inputs;
 		std::vector<Connector*> outputs;
 		bool sign;
-		double delay;
+		int delay;
 		std::string outType;
 		int timeASAP;
 		int timeALAP;
